+++ django_rubble/admin/__init__.py
@@ -1,0 +1,11 @@
+from .history_admin import HistoryAdmin, HistoryTabularInline
+from .number_admin import NumberAdmin
+from .stamped_admin import StampedAdmin, StampedTabularInline
+
+__all__ = [
+    "HistoryAdmin",
+    "HistoryTabularInline",
+    "NumberAdmin",
+    "StampedAdmin",
+    "StampedTabularInline",
+]